+++ bindings/python/examples/how_tos/simple_transaction/simple_transaction.py
@@ -18,10 +18,10 @@ if 'STRONGHOLD_PASSWORD' not in os.environ:
 
 wallet.set_stronghold_password(os.environ["STRONGHOLD_PASSWORD"])
 
-outputs = [{
+params = [{
     "address": "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu",
     "amount": "1000000",
 }]
 
-transaction = account.send_amount(outputs)
-print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction["blockId"]}')
+transaction = account.send(outputs)
+print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction.blockId}')