+++ sdk/tests/client/node_api.rs
@@ -6,7 +6,7 @@
 use iota_sdk::{
     client::{
         api::GetAddressesOptions, bech32_to_hex, node_api::indexer::query_parameters::QueryParameter,
-        request_funds_from_faucet, secret::SecretManager, Client,
+        request_funds_from_faucet, secret::SecretManager, Client, NodeInfoWrapper,
     },
     types::block::{
         address::ToBech32Ext,
@@ -401,4 +401,20 @@ async fn test_mqtt() {
         }
     }
     client.subscriber().disconnect().await.unwrap();
+}
+
+#[tokio::test]
+async fn test_call_plugin_route() {
+    let c = setup_client_with_node_health_ignored().await;
+
+    // we call the "custom" plugin "node info"
+    let plugin_res: NodeInfoWrapper = c
+        .call_plugin_route("api/core/v2/", "GET", "info", vec![], None)
+        .await
+        .unwrap();
+
+    let info = c.get_info().await.unwrap();
+
+    // Just check name as info can change between 2 calls
+    assert_eq!(plugin_res.node_info.name, info.node_info.name);
 }